moved scheduleNextDisplay to VotingState
