+++ src/main/java/org/braekpo1nt/mctmanager/games/gamemanager/states/event/VotingState.java
@@ -1,6 +1,8 @@
 package org.braekpo1nt.mctmanager.games.gamemanager.states.event;
 
+import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.format.NamedTextColor;
 import org.braekpo1nt.mctmanager.commands.manager.commandresult.CommandResult;
 import org.braekpo1nt.mctmanager.games.game.enums.GameType;
 import org.braekpo1nt.mctmanager.games.gamemanager.GameManager;
@@ -9,20 +11,27 @@ import org.braekpo1nt.mctmanager.games.gamemanager.event.EventData;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.ContextReference;
 import org.braekpo1nt.mctmanager.games.gamemanager.states.event.delay.StartingGameDelayState;
 import org.braekpo1nt.mctmanager.games.voting.VoteManager;
+import org.braekpo1nt.mctmanager.participant.Participant;
+import org.braekpo1nt.mctmanager.ui.UIUtils;
 import org.braekpo1nt.mctmanager.ui.timer.Timer;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
+import org.bukkit.scheduler.BukkitRunnable;
+import org.bukkit.scheduler.BukkitTask;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Random;
 import java.util.Set;
 
 public class VotingState extends EventState {
     
     private final VoteManager voteManager;
     private final Timer timer;
+    private @Nullable BukkitTask display;
     
     public VotingState(@NotNull GameManager context, @NotNull ContextReference contextReference, @NotNull EventData eventData) {
         super(context, contextReference, eventData);
@@ -57,15 +66,91 @@ public class VotingState extends EventState {
     public void exit() {
         voteManager.cancelVote();
         timer.cancel();
+        if (display != null) {
+            display.cancel();
+        }
     }
     
-    protected void onVoteExecuted(GameType gameType, String configFile) {
+    protected void onVoteExecuted(List<GameType> gameTypes, String configFile) {
         timer.cancel();
-        String chosenConfigFile = eventData.getConfig().getGameConfigs().getOrDefault(gameType, configFile);
-        eventData.getConfig().isWeightedVoting();
-        context.setState(new StartingGameDelayState(
-                context, contextReference, eventData,
-                gameType, chosenConfigFile));
+        if (eventData.getConfig().isWeightedVoting()) {
+            Random random = new Random();
+            scheduleNextDisplay(gameTypes, 5, random, true);
+        } else {
+            GameType gameType = gameTypes.getFirst();
+            String chosenConfigFile = eventData.getConfig().getGameConfigs().getOrDefault(gameType, configFile);
+            context.setState(new StartingGameDelayState(
+                    context, contextReference, eventData,
+                    gameType, chosenConfigFile));
+        }
+    }
+    
+    public void scheduleNextDisplay(List<GameType> votes, final long numberOfTicks, Random random, final boolean displayIsRed) {
+        
+        this.display = new BukkitRunnable() {
+            @Override
+            public void run() {
+                boolean redTitle;
+                int selectedInt;
+                GameType gameType;
+                long nextNumberOfTicks;
+                if (votes.size() > 1) {
+                    selectedInt = random.nextInt(votes.size());
+                    gameType = votes.get(selectedInt);
+                    votes.remove(selectedInt);
+                    if (votes.size() > 6) {
+                        nextNumberOfTicks = 5;
+                    } else {
+                        nextNumberOfTicks = switch (votes.size()) {
+                            case 6 -> 7;
+                            case 5 -> 9;
+                            case 4 -> 13;
+                            case 3 -> 16;
+                            default -> 20;
+                        };
+                    }
+                    if (displayIsRed) {
+                        Audience.audience( // Use this for display, modify color
+                                Audience.audience(onlineParticipants.values()),
+                                Audience.audience(onlineAdmins)
+                        ).showTitle(UIUtils.defaultTitle(
+                                Component.empty()
+                                        .append(Component.text(gameType.getTitle()))
+                                        .color(NamedTextColor.RED),
+                                Component.empty()
+                        ));
+                        redTitle = false;
+                    } else {
+                        Audience.audience( // Use this for display, modify color
+                                Audience.audience(onlineParticipants.values()),
+                                Audience.audience(onlineAdmins)
+                        ).showTitle(UIUtils.defaultTitle(
+                                Component.empty()
+                                        .append(Component.text(gameType.getTitle()))
+                                        .color(NamedTextColor.YELLOW),
+                                Component.empty()
+                        ));
+                        redTitle = true;
+                    }
+                } else {
+                    gameType = votes.getFirst();
+                    Audience.audience( // Use this for display, modify color
+                            Audience.audience(onlineParticipants.values()),
+                            Audience.audience(onlineAdmins)
+                    ).showTitle(UIUtils.defaultTitle(
+                            Component.empty()
+                                    .append(Component.text(gameType.getTitle()))
+                                    .color(NamedTextColor.BLUE),
+                            Component.empty()
+                    ));
+                    redTitle = true;
+                    nextNumberOfTicks = 0L;
+                }
+                if (!votes.isEmpty()) {
+                    scheduleNextDisplay(votes, nextNumberOfTicks, random, redTitle);
+                }
+            }
+        }.runTaskLater(plugin, numberOfTicks);
     }
     
     @Override
